+++ scripts/select_speaker.py
